code cleanup and modernization
- remove a lot of unnecessary std::move()
- use range-based algorithms
- use views when iterating over map
- fix many typos
- other minor fixes detected by static analysis

+++ plugins/countdown_timer/impl/countdown_timer.cpp
@@ -37,7 +37,7 @@ void CountdownTimer::stop()
   emit activityChanged(is_active_);
 }
 
-void CountdownTimer::setInterval(int seconds)
+void CountdownTimer::setInterval(qint64 seconds)
 {
   setTimeLeft(seconds);
   interval_ = seconds;
